+++ ros2_ws/src/tactile_vision/tactile_vision/__init__.py
@@ -1,0 +1,2 @@
+"""tactile_vision package."""
+